+++ integration-service/src/main/java/com/Gritty/Linki/domain/oAuth/account/repository/FindIdRepository.java
@@ -1,4 +1,4 @@
-package com.Gritty.Linki.domain.user.User.repository;
+package com.Gritty.Linki.domain.oAuth.account.repository;
 
 import com.Gritty.Linki.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
