+++ integration-service/src/main/java/com/Gritty/Linki/domain/oAuth/dto/FindIdResponseDto.java
@@ -1,15 +1,18 @@
-package com.Gritty.Linki.domain.user.User.dto;
+package com.Gritty.Linki.domain.oAuth.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class FindIdRequestDto {
+@Builder
+public class FindIdResponseDto {
+    private String userId;
     private String userName;
     private String userEmail;
 }
