+++ integration-service/src/main/java/com/Gritty/Linki/domain/oAuth/dto/VerificationRequestDto.java
@@ -1,18 +1,16 @@
-package com.Gritty.Linki.domain.user.User.dto;
+package com.Gritty.Linki.domain.oAuth.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class FindIdResponseDto {
-    private String userId;
+public class VerificationRequestDto {
     private String userName;
     private String userEmail;
+    private String verificationCode;
 }
